requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
starlette~=0.27.0
SQLAlchemy==1.4.39


#Google 
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-oauth2-tool==0.0.3
googleapis-common-protos==1.61.0
google-oauth2-tool==0.0.3
SQLAlchemy~=2.0.22
bcrypt~=4.0.1
PyJWT~=2.8.0